api_hackathon_march_2024/Techiee Hackers/requirements.txt
requests
python-dotenv
openai
streamlit
os